# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.DocumentListResponse:
    properties:
      count:
        type: integer
      documents: {}
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
...
  handlers.StatusResponse:
    properties:
      document_id:
        type: string
      status:
        type: string
    type: object
  handlers.SuccessResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  repositories.Chunk:
    properties:
      chunk_index:
        type: integer
      document_id:
        type: string
      embedding:
        items:
          type: number
        type: array
      id:
        type: string
      metadata:
        additionalProperties: true
        type: object
      page_number:
        type: integer
      text:
        type: string
      token_count:
        type: integer
    type: object
  repositories.Document:
    properties:
      chunk_count:
        type: integer
      chunk_overlap:
        type: integer
      chunk_size:
        type: integer
      chunking_strategy:
        description: Processing metadata
        type: string
      collection:
        type: string
      created_at:
        type: string
      document_id:
        type: string
      extract_metadata:
        type: boolean
      file_size:
        type: integer
      filename:
        type: string
      llm_model:
        type: string
      llm_provider:
        description: LLM metadata
        type: string
      max_pages:
        type: integer
      metadata:
        additionalProperties: true
        type: object
      num_questions:
        type: integer
      status:
        $ref: '#/definitions/repositories.DocumentStatus'
      stored_in_vector_db:
        type: boolean
      updated_at:
        type: string
    type: object
  repositories.DocumentStatus:
    enum:
    - pending
    - processing
    - completed
    - failed
    - deleted
    type: string
    x-enum-varnames:
    - DocumentStatusPending
    - DocumentStatusProcessing
    - DocumentStatusCompleted
    - DocumentStatusFailed
    - DocumentStatusDeleted
...
  services.GetDocumentChunksResponse:
    properties:
      chunks:
        items:
          $ref: '#/definitions/repositories.Chunk'
        type: array
      collection:
        type: string
      document_id:
        type: string
      limit:
        type: integer
      offset:
        type: integer
      total_count:
        type: integer
    type: object
...
  services.UploadDocumentResponse:
    properties:
      chunk_count:
        type: integer
      collection:
        type: string
      document_id:
        type: string
      filename:
        type: string
      job_id:
        type: string
      message:
        type: string
      metadata:
        additionalProperties: true
        type: object
      processing_time_ms:
        type: number
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive API for document processing, analysis, and risk assessment
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Risk Analyzer API
  version: "1.0"
paths:
...
  /api/v1/documents:
    get:
      description: Get a list of all documents
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DocumentListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List documents
      tags:
      - documents
  /api/v1/documents/{id}:
    delete:
      description: Delete a document and all its chunks
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete document
      tags:
      - documents
    get:
      description: Get document by ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repositories.Document'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get document
      tags:
      - documents
  /api/v1/documents/{id}/chunks:
    get:
      description: Get all chunks for a specific document from the vector store
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Collection name (optional, will use document's collection if
          not provided)
        in: query
        name: collection
        type: string
      - default: 100
        description: Limit results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.GetDocumentChunksResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get document chunks
      tags:
      - documents
  /api/v1/documents/{id}/status:
    get:
      description: Get the processing status of a document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get document status
      tags:
      - documents
  /api/v1/documents/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload and process a document for vector storage
      parameters:
      - description: Document file
        in: formData
        name: file
        required: true
        type: file
      - description: Collection name
        in: formData
        name: collection
        required: true
        type: string
      - default: semantic
        description: Chunking strategy
        in: formData
        name: chunking_strategy
        type: string
      - default: 512
        description: Chunk size
        in: formData
        name: chunk_size
        type: integer
      - default: 50
        description: Chunk overlap
        in: formData
        name: chunk_overlap
        type: integer
      - default: false
        description: Extract metadata
        in: formData
        name: extract_metadata
        type: boolean
      - default: 3
        description: Number of questions
        in: formData
        name: num_questions
        type: integer
      - default: 0
        description: Max pages to process
        in: formData
        name: max_pages
        type: integer
      - default: false
        description: Process asynchronously
        in: formData
        name: async
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.UploadDocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Upload a document
      tags:
      - documents
...
  /documents/upload:
    post:
      consumes:
      - multipart/form-data
      deprecated: true
      description: 'DEPRECATED: Use /api/v1/documents/upload instead. Upload a PDF
        or document file and process it'
      parameters:
      - description: Document file to upload
        in: formData
        name: file
        required: true
        type: file
      - default: sentence
        description: Chunking strategy
        in: query
        name: chunking_strategy
        type: string
      - default: 512
        description: Chunk size
        in: query
        name: chunk_size
        type: integer
      - default: 50
        description: Chunk overlap
        in: query
        name: chunk_overlap
        type: integer
      - default: false
        description: Store in vector DB
        in: query
        name: store_in_vector_db
        type: boolean
      - default: documents
        description: Collection name
        in: query
        name: collection_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: '[DEPRECATED] Upload and process document'
      tags:
      - deprecated
...
swagger: "2.0"